Restructured each cache to be limited based on amount of memory in use instead of number of objects in the cache. Added new options for limiting caches by a set memory size. Functionality added for keeping track of the total size of objects in each cache.

+++ earth_enterprise/src/common/MemoryMonitor.cpp
@@ -55,7 +55,7 @@ void CalculateMemoryUsage(void) {
         memoryUsed = 0;
     }
     else {
-        memoryUsed = (((static_cast<float> (std::accumulate(memoryVars.begin()+1, memoryVars.end(), memoryVars[0], std::minus<ulong>())))/memoryVars[0])*100);
+        memoryUsed = static_cast<uint> (((static_cast<float> (std::accumulate(memoryVars.begin()+1, memoryVars.end(), memoryVars[0], std::minus<ulong>())))/memoryVars[0])*100);
     }
     WriteToMemFile(memoryUsed);
 }
